requirements.txt
matplotlib==3.1.1
numpy==1.17.2
six==1.12.0
tensorflow==1.15.0
Pillow==8.3.2
scikit_learn==0.24.2
